@@ -10,8 +10,8 @@ public class MyApp extends Application {
         super.onCreate();
         // if you are using push module make sure to call SegmentifyManager.Instance.setPushConfig first
         SegmentifyManager.INSTANCE.setPushConfig("https://gimli-qa.segmentify.com");
-        SegmentifyManager.INSTANCE.config(this, "e4dfb71f-eac8-44b7-8938-ce9cce910cb2","http://gandalf-qa.segmentify.com","segmentify-search-showcase.mybigcommerce.com");
-        SegmentifyManager.INSTANCE.setSessionKeepSecond(100000);
+        SegmentifyManager.INSTANCE.config(this, "5c571072-068e-40c5-8dbc-d8448158de19","https://gandalf-qa.segmentify.com","demo.segmentify.com");
+        SegmentifyManager.INSTANCE.setSessionKeepSecond(86400);
         SegmentifyManager.INSTANCE.logStatus(true);
     }
 }
