@@ -25,7 +25,7 @@ object SegmentifyManager {
     private fun addClickedBanner(banner: BannerOperationsModel) {
 
         clickedBanners.forEach {
-            if (    it.group.equals(banner.group) &&
+            if (it.group.equals(banner.group) &&
                     it.title.equals(banner.title) &&
                     it.order!!.equals(banner.order)) {
                 return
@@ -49,7 +49,7 @@ object SegmentifyManager {
         clientPreferences?.setSessionKeepSeconds(sessionKeepSecond)
     }
 
-    fun setConfig(apiKey: String, dataCenterUrl: String, subDomain: String ) {
+    fun setConfig(apiKey: String, dataCenterUrl: String, subDomain: String) {
         this.configModel.apiKey = apiKey
         this.configModel.dataCenterUrl = dataCenterUrl
         this.configModel.subDomain = subDomain
@@ -57,7 +57,7 @@ object SegmentifyManager {
         ConnectionManager.rebuildServices()
     }
 
-    fun setPushConfig(dataCenterUrlPush: String ) {
+    fun setPushConfig(dataCenterUrlPush: String) {
         this.configModel.dataCenterUrlPush = dataCenterUrlPush
     }
 
@@ -67,7 +67,7 @@ object SegmentifyManager {
 
     fun config(context: Context, appKey: String, dataCenterUrl: String, subDomain: String) {
 
-        if (appKey.isNullOrBlank() || dataCenterUrl.isNullOrBlank() || subDomain.isNullOrBlank()) {
+        if (appKey.isBlank() || dataCenterUrl.isBlank() || subDomain.isBlank()) {
             SegmentifyLogger.printErrorLog("Api is not initialized, you can not enter null or empty parameter to config, please recheck your config parameters")
             return
         }
@@ -110,9 +110,9 @@ object SegmentifyManager {
         })
     }
 
-    fun sendSearchPageView(pageModel: SearchPageModel, segmentifyCallback: SegmentifyCallback<SearchResponseModel>){
+    fun sendSearchPageView(pageModel: SearchPageModel, segmentifyCallback: SegmentifyCallback<SearchResponseModel>) {
         pageModel.eventName = Constant.searchViewEventName
-        EventController.sendSearchView(pageModel, object : SegmentifyCallback<SearchResponseModel>{
+        EventController.sendSearchView(pageModel, object : SegmentifyCallback<SearchResponseModel> {
             override fun onDataLoaded(data: SearchResponseModel) {
                 segmentifyCallback.onDataLoaded(data)
             }
@@ -255,8 +255,7 @@ object SegmentifyManager {
     fun sendProductView(productId: String, title: String, categories: ArrayList<String>, price: Double,
                         brand: String?, stock: Boolean?, url: String, image: String, imageXS: String?,
                         imageS: String?, imageM: String?, imageL: String?, imageXL: String?, gender: String?,
-                        colors: ArrayList<String>?, sizes: ArrayList<String>?, labels: ArrayList<String>?, noUpdate: Boolean?
-                        , segmentifyCallback: SegmentifyCallback<ArrayList<RecommendationModel>>) {
+                        colors: ArrayList<String>?, sizes: ArrayList<String>?, labels: ArrayList<String>?, noUpdate: Boolean?, segmentifyCallback: SegmentifyCallback<ArrayList<RecommendationModel>>) {
         var productModel = ProductModel()
         productModel.eventName = Constant.productViewEventName
 
@@ -547,12 +546,44 @@ object SegmentifyManager {
 
     }
 
+    fun sendImpression(interactionModel: InteractionModel) {
+        if (interactionModel.instanceId.isNullOrBlank()) {
+            SegmentifyLogger.printErrorLog("You must fill instanceId before sending interaction event")
+            return
+        }
+        if (interactionModel.interactionId.isNullOrBlank()) {
+            SegmentifyLogger.printErrorLog("You must fill interactionId before sending interaction event")
+            return
+        }
+
+        interactionModel.eventName = Constant.interactionEventName
+        interactionModel.type = Constant.impressionStep
+
+        EventController.sendInteractionEvent(interactionModel)
+    }
+
     fun sendWidgetView(instanceId: String, interactionId: String) {
         var interactionModel = InteractionModel()
         interactionModel.eventName = Constant.interactionEventName
         interactionModel.type = Constant.widgetViewStep
         interactionModel.instanceId = instanceId
         interactionModel.interactionId = interactionId
+
+        EventController.sendInteractionEvent(interactionModel)
+    }
+
+    fun sendWidgetView(interactionModel: InteractionModel) {
+        if (interactionModel.instanceId.isNullOrBlank()) {
+            SegmentifyLogger.printErrorLog("You must fill instanceId before sending interaction event")
+            return
+        }
+        if (interactionModel.interactionId.isNullOrBlank()) {
+            SegmentifyLogger.printErrorLog("You must fill interactionId before sending interaction event")
+            return
+        }
+
+        interactionModel.eventName = Constant.interactionEventName
+        interactionModel.type = Constant.widgetViewStep
 
         EventController.sendInteractionEvent(interactionModel)
     }
@@ -567,12 +598,44 @@ object SegmentifyManager {
         EventController.sendInteractionEvent(interactionModel)
     }
 
-    fun sendSearchClickView(instanceId: String, interactionId: String){
+    fun sendClickView(interactionModel: InteractionModel) {
+        if (interactionModel.instanceId.isNullOrBlank()) {
+            SegmentifyLogger.printErrorLog("You must fill instanceId before sending interaction event")
+            return
+        }
+        if (interactionModel.interactionId.isNullOrBlank()) {
+            SegmentifyLogger.printErrorLog("You must fill interactionId before sending interaction event")
+            return
+        }
+
+        interactionModel.eventName = Constant.interactionEventName
+        interactionModel.type = Constant.clickStep
+
+        EventController.sendInteractionEvent(interactionModel)
+    }
+
+    fun sendSearchClickView(instanceId: String, interactionId: String) {
         var interactionModel = InteractionModel()
         interactionModel.eventName = Constant.interactionEventName
         interactionModel.type = Constant.searchStep
         interactionModel.instanceId = instanceId
         interactionModel.interactionId = interactionId
+
+        EventController.sendInteractionEvent(interactionModel)
+    }
+
+    fun sendSearchClickView(interactionModel: InteractionModel) {
+        if (interactionModel.instanceId.isNullOrBlank()) {
+            SegmentifyLogger.printErrorLog("You must fill instanceId before sending interaction event")
+            return
+        }
+        if (interactionModel.interactionId.isNullOrBlank()) {
+            SegmentifyLogger.printErrorLog("You must fill interactionId before sending interaction event")
+            return
+        }
+
+        interactionModel.eventName = Constant.interactionEventName
+        interactionModel.type = Constant.searchStep
 
         EventController.sendInteractionEvent(interactionModel)
     }
@@ -732,11 +795,10 @@ object SegmentifyManager {
         }
 
         if (notificationModel.type == NotificationType.CLICK) {
-            if (notificationModel.instanceId.isNullOrEmpty() && notificationModel.productId.isNullOrEmpty() ) {
+            if (notificationModel.instanceId.isNullOrEmpty() && notificationModel.productId.isNullOrEmpty()) {
                 SegmentifyLogger.printErrorLog("You must fill deviceToken before accessing notification click event")
                 return
-            }
-            else{
+            } else {
                 clientPreferences?.setPushCampaignId(notificationModel.instanceId!!)
                 clientPreferences?.setPushCampaignProductId(notificationModel.productId!!)
                 sendClickView(notificationModel.instanceId!!, notificationModel.productId!!);
@@ -746,7 +808,23 @@ object SegmentifyManager {
     }
 
     fun getTrackingParameters(): UtmModel {
-        return  UtmModel()
+        return UtmModel()
+    }
+
+    fun clearLastSearchHistory(userModel: UserModel) {
+        userModel.eventName = Constant.userOperationEventName
+        userModel.userOperationStep = Constant.deleteLastSearchStep
+
+        EventController.sendUserOperation(userModel)
+    }
+
+    fun clearLastSearchHistory(lastSearchDeletedKeyword: String?) {
+        var userModel = UserModel()
+        userModel.eventName = Constant.userOperationEventName
+        userModel.userOperationStep = Constant.deleteLastSearchStep
+        userModel.lastSearchDeletedKeywords = lastSearchDeletedKeyword
+
+        EventController.sendUserOperation(userModel)
     }
 
 }
