@@ -1,22 +1,17 @@
 package com.segmentify.segmentifyandroidsdk.utils
 
 import android.content.Context
-import java.util.*
 
-/**
- * Created by rifatmertdolar on 11/04/2018.
- */
-
-class ClientPreferences(context : Context) : PreferencesManager(context){
+class ClientPreferences(context: Context) : PreferencesManager(context) {
 
 
-    private  val USER_NAME = "USERNAME"
-    private  val EMAIL = "EMAIL"
+    private val USER_NAME = "USERNAME"
+    private val EMAIL = "EMAIL"
     private val USER_ID = "USER_ID"
     private val SESSION_ID = "SESSION_ID"
     private val IS_LOG_VISIBLE = "IS_LOG_VISIBLE"
     private val API_URL = "API_URL"
-    private val SAVED_DATE_FOR_SESSION = "SAVED_DATE_FOR_SESSION"
+    private val SAVED_DATE_FOR_SESSION_MILLIS = "SAVED_DATE_FOR_SESSION_MILLIS"
     private val SESSION_KEEP_SECONDS = "SESSION_KEEP_SECONDS"
     private val PUSH_CAMPAIGN_ID = "PUSH_CAMPAIGN_ID"
     private val PUSH_CAMPAIGN_PRODUCT_ID = "PUSH_CAMPAIGN_PRODUCT_ID"
@@ -53,12 +48,12 @@ class ClientPreferences(context : Context) : PreferencesManager(context){
         putBoolean(IS_LOG_VISIBLE, isLogVisible)
     }
 
-    fun getSavedDateForSession(): Date? {
-        return getObject(SAVED_DATE_FOR_SESSION, Date::class.java)
+    fun getSavedDateForSession(): Long? {
+        return getLong(SAVED_DATE_FOR_SESSION_MILLIS)
     }
 
-    fun setSavedDateforSession(date: Date) {
-        putObject(SAVED_DATE_FOR_SESSION, date)
+    fun setSavedDateforSession(date: Long) {
+        putLong(SAVED_DATE_FOR_SESSION_MILLIS, date)
     }
 
     fun getSessionKeepSeconds(): Int {
@@ -70,18 +65,19 @@ class ClientPreferences(context : Context) : PreferencesManager(context){
     }
 
     fun setEmail(email: String) {
-        putString(EMAIL,email)
+        putString(EMAIL, email)
     }
 
     fun setUserName(userName: String) {
-        putString(USER_NAME,userName)
+        putString(USER_NAME, userName)
     }
 
-    fun setPushCampaignId(instanceId: String){
-        putString(PUSH_CAMPAIGN_ID,instanceId)
+    fun setPushCampaignId(instanceId: String) {
+        putString(PUSH_CAMPAIGN_ID, instanceId)
     }
-    fun setPushCampaignProductId(productId: String){
-        putString(PUSH_CAMPAIGN_PRODUCT_ID,productId)
+
+    fun setPushCampaignProductId(productId: String) {
+        putString(PUSH_CAMPAIGN_PRODUCT_ID, productId)
     }
 
 }
