@@ -2,12 +2,9 @@ package com.segmentify.segmentifyandroidsdk.utils
 
 import android.content.Context
 import android.content.SharedPreferences
-import java.text.SimpleDateFormat
 import com.google.gson.Gson
-import java.text.ParseException
-import java.util.*
 
-abstract class PreferencesManager(targetContext: Context){
+abstract class PreferencesManager(targetContext: Context) {
     private val defaultPackageName = "AppcentPackageSettings"
     private val emptyString = ""
     private var preferenceName: String = ""
@@ -17,10 +14,6 @@ abstract class PreferencesManager(targetContext: Context){
     init {
         targetPreferences = targetContext.getSharedPreferences(preferenceName, Context.MODE_PRIVATE)
     }
-
-    /*constructor(targetContext : Context,preferenceName : String) : this(targetContext) {
-
-    }*/
 
     fun clearKey(key: String) {
         if (targetPreferences.contains(key)) {
@@ -50,43 +43,6 @@ abstract class PreferencesManager(targetContext: Context){
 
     fun getLong(key: String): Long {
         return targetPreferences.getLong(key, 0)
-    }
-
-    fun <T> getObject(key: String, targetType: Class<T>): T? {
-        return getObject(key, targetType, null)
-    }
-
-    fun <T> getObject(key: String, targetType: Class<T>, defaultValue: T?): T? {
-        if (targetPreferences.contains(key)) {
-            var preferenceTarget = targetPreferences.getString(key, "")
-            if (preferenceTarget != "") {
-                if(preferenceTarget.contains("AM") || preferenceTarget.contains("PM")){
-                    try{
-                        return gson.fromJson(preferenceTarget, targetType)
-                    }
-                    catch (e:Exception){
-                        e.printStackTrace()
-                    }
-                }
-                else{
-                    val df = SimpleDateFormat("MMM dd, yyyy hh:mm:ss")
-                    val outputformat = SimpleDateFormat("MMM dd, yyyy hh:mm:ss aa")
-                    var date: Date? = null
-                    var output: String? = null
-
-                    try {
-                        date = df.parse(preferenceTarget)
-                        output = '"'+outputformat.format(date).toString()+'"'
-
-                    } catch (pe: ParseException) {
-                        pe.printStackTrace()
-                    }
-                    return gson.fromJson(output, targetType)
-                }
-
-            }
-        }
-        return defaultValue
     }
 
     fun getPreferenceName(): String {
